2D biofilm simulation
Overview
--------

The following implementation is based on the outline in chapter 10.5 of
Introduction to Computational Science by Angela and George Shiflet. This
implementation also takes inspiration from Data-Oriented Design and the
Entity-Component-System (ECS) pattern.

The simulation consists of a bacteria grid and a nutrients grid that update in
tandem. Nutrients continuously diffuse into the simulation area and the bacteria
consume the nutrients while they grow through division. This generates the
various structures seen in the animated image. If the nutrients run out in a
particular cell, then the bacterium dies. The total area reached by the bacteria
depend on parameters such as the growth rate, the diffusion rate, the
consumption rate, etc.

Implementation
--------------

* Components: cells in the simulation are represented as one-dimensional arrays.
  There are several arrays that exist in parallel to one another. Two are the
  "real" versions of the bacteria and nutrients grid, and two are copies with
  buffer cells wrapped around them.

* Entities: Manages the list of components. A bit-field is used to toggle
  different components on and off, which defines the difference between cells
  and the padded region (ghost cells).

* Systems: The algorithms that update the components.

* Others: Utility functions and neighbor finder and picker functions are stored
  outside this framework.

* Benchmarking: A simulation timer is directly built into the source code. The
  results can be saved to a csv file via a command-line option.

Scripts
-------

Two additional scripts are included with this release. The first is make.sh,
which is a wrapper script for building the project using cmake. The second
script is animate.py, which generates animations using the simulation history
files.

Documentation
-------------

A full README file is included that gives an overview of the biofilm simulation
and provides a link to Shiflet and Shiflet's module write-up. The README
provides basic explainers on how to compile the code, use the command-line
interface, and how to install the development environment. An annotated visual
layout of the files and directories is also provided.

License
-------

This project is released under the MIT license.