Source code documentation
Summary
-------

Documentation has been added for all of the functions and structs used in the
simulation. The documentation is in the Doxygen format. Function and struct
documentation has been added to the following source files,

* components.hpp
* consumption_system.cpp
* copy_system.cpp
* data.hpp
* diffusion_system.cpp
* entities.hpp
* growth_system.cpp
* initializers.cpp
* neighbors.cpp
* simulation.cpp
* utilities.cpp

In addition to the standard form of documentation, further annotations have also
been included to further clarify certain aspects of the simulation. Those
additional annotations have been added to the following files,

* consumption_system.cpp
* diffusion_system.cpp
* growth_system.cpp
* simulation.cpp

The UnpaddedToPaddedIDConverter function object was removed during this update.
It's intended utility was superseded by the locations and flags components.